requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
aiofiles>=23.2.1
aiohttp>=3.8.6
asyncio
requests>=2.31.0
pillow>=10.0.1
youtube-dl>=2021.12.17
yt-dlp>=2023.9.24
psutil>=5.9.6
speedtest-cli>=2.1.3
googletrans>=4.0.0
wikipedia>=1.4.0
pytz>=2023.3
python-dotenv>=1.0.0
motor>=3.3.2
pymongo>=4.6.0
redis>=5.0.1
sqlalchemy>=2.0.23
cryptography>=41.0.7
qrcode>=7.4.2
opencv-python>=4.8.1.78
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0
pandas>=2.1.4
scikit-learn>=1.3.2
textblob>=0.17.1
nltk>=3.8.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2
fake-useragent>=1.4.0
python-telegram-bot>=20.7
telethon>=1.34.0
pyrogram-mod>=2.0.106
asyncpg>=0.29.0
aiogram>=3.2.0
python-magic>=0.4.27
ffmpeg-python>=0.2.0
gtts>=2.4.0
pydub>=0.25.1
mutagen>=1.47.0
python-barcode>=0.15.1
reportlab>=4.0.7
fpdf2>=2.7.6
openpyxl>=3.1.2
xlsxwriter>=3.1.9
python-docx>=1.1.0
markdown>=3.5.1
html2text>=2020.1.16
python-dateutil>=2.8.2
humanize>=4.8.0
colorama>=0.4.6
rich>=13.7.0
tabulate>=0.9.0
progress>=1.6
tqdm>=4.66.1
click>=8.1.7
typer>=0.9.0
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
starlette>=0.27.0
jinja2>=3.1.2
aioredis>=2.0.1
celery>=5.3.4
kombu>=5.3.4
billiard>=4.2.0
vine>=5.1.0
amqp>=5.2.0